chore(opentelemetry source): Move tests into separate module
